+++ file_type_icon.py
@@ -17,7 +17,6 @@ from .src.zukan_icon_theme.commands.commands import (  # noqa: E402
     DeletePreference,  # noqa: F401
     DeletePreferences,  # noqa: F401
     DeleteSyntax,  # noqa: F401
-    DeleteSyntaxes,  # noqa: F401
     DeleteTheme,  # noqa: F401
     DeleteThemes,  # noqa: F401
     InstallPreference,  # noqa: F401
@@ -27,6 +26,12 @@ from .src.zukan_icon_theme.commands.commands import (  # noqa: E402
     RebuildFiles,  # noqa: F401
     RebuildPreferences,  # noqa: F401
     RebuildSyntaxes,  # noqa: F401
+)
+from .src.zukan_icon_theme.commands.commands_settings import (  # noqa: E402
+    DisableIcon,  # noqa: F401
+    DisableTheme,  # noqa: F401
+    EnableIcon,  # noqa: F401
+    EnableTheme,  # noqa: F401
 )
 from .src.zukan_icon_theme.events.install import InstallEvent  # noqa: E402
 from .src.zukan_icon_theme.events.settings import SettingsEvent  # noqa: E402