+++ src/zukan_icon_theme/helpers/load_save_settings.py
@@ -13,3 +13,16 @@ def get_settings(file_settings: str, option: str = None):
         return sublime.load_settings(file_settings).get(option)
     if option is None:
         return sublime.load_settings(file_settings)
+
+
+def set_save_settings(file_settings: str, option: str, value: list):
+    """
+    Modify and save settings options.
+
+    Parameters:
+    file_settings (str) -- sublime-settings file.
+    option (str) -- set option key.
+    value (list) --  option vslue.
+    """
+    sublime.load_settings(file_settings).set(option, value)
+    sublime.save_settings(file_settings)